profiles

+++ uu_tatranci_maing01-hi/src/bricks/skolsky-bufet-ready.js
@@ -68,15 +68,22 @@ export const SkolskyBufetReady = createVisualComponent({
         },
         {
           cell: (cellProps) => {
-            return (
-              <>
-                <UU5.Bricks.Button onClick={() => handleRemove(cellProps)} className={CLASS_NAMES.buttons()}> Zmazať produkt </UU5.Bricks.Button> <br/> <br/>
-                <UU5.Bricks.Modal ref={modalRef}/>
-                <UU5.Bricks.Button onClick={() => handleUpdate(cellProps.data)} className={CLASS_NAMES.buttons()}> Aktualizovať produkt </UU5.Bricks.Button> <br/> <br/>
-                <UU5.Bricks.Button onClick={() => handleDetail(cellProps.data)} className={CLASS_NAMES.buttons()}> Detail </UU5.Bricks.Button> <br/> <br/>
-                <UU5.Bricks.Button onClick={() => handleReservation(cellProps)} className={CLASS_NAMES.buttons()}> Rezervovať </UU5.Bricks.Button>
-              </>
-            )
+            if (cellProps.data.profileList.includes("Authorities")) {
+              return (
+                <>
+                  <UU5.Bricks.Button onClick={() => handleRemove(cellProps)} className={CLASS_NAMES.buttons()}> Zmazať produkt </UU5.Bricks.Button> <br/> <br/>
+                  <UU5.Bricks.Button onClick={() => handleUpdate(cellProps.data)} className={CLASS_NAMES.buttons()}> Aktualizovať produkt </UU5.Bricks.Button> <br/> <br/>
+                </>
+              )
+            } else {
+              return (
+                <>
+                  <UU5.Bricks.Modal ref={modalRef}/>
+                  <UU5.Bricks.Button onClick={() => handleDetail(cellProps.data)} className={CLASS_NAMES.buttons()}> Detail </UU5.Bricks.Button> <br/> <br/>
+                  <UU5.Bricks.Button onClick={() => handleReservation(cellProps)} className={CLASS_NAMES.buttons()}> Rezervovať </UU5.Bricks.Button>
+                </>
+              )
+            }
           },
         }
       ];
@@ -103,7 +110,7 @@ export const SkolskyBufetReady = createVisualComponent({
     function handleDetail(data) {
       modalRef.current.open({
         header: " ",
-        content: (<SkolskyBufetReady data={data} handleClose={handleClose} />),
+        content: (<SkolskyBufetReady data={data} handleClose={handleClose}/>),
       })
       console.log(data);
